fix: extracted types
